vimrc: Change behavior of (normal) j/k
Generally, j/k will make you jump up/down lines, placing your cursor
in the same column, with the row de/incremented.  However, this is
unintuitive in the case where you're moving into a line that wraps.
For example, if we were editing two lines like (where the current
cursor position is the _ char) ...

    +-------------------+
    |1| abcdefghijklmnop|
    | | qrstuvwxyz      |
    |2| 0123_56789      |
    +-------------------|

... and then we pressed `k`, we would have ...

    +-------------------+
    |1| abcd_fghijklmnop|
    | | qrstuvwxyz      |
    |2| 0123456789      |
    +-------------------|

... although it would be more intuitive if the cursor only moved up
a "vertical" position as determined *visually*, like ...

    +-------------------+
    |1| abcdefghijklmnop|
    | | qrst_vwxyz      |
    |2| 0123456789      |
    +-------------------|

So that's what this patch implements :^)