bashrc: Remove "cat" alias
This alias was so that "unprintable" characters would be displayed
by cat in the same way that Vim displays them: as ^escapes.

However, this has the unfortunate side effect of not printing things
out *exactly* as they appear in source, which is not particularly
desirable.  In fact, I found myself often invoking cat by /bin/cat.

I think it's more sane to leave the ^escape behavior behind the -v
flag :^)