git-nav: Add functionality for moving thru recent checkout history
This functionality is extremely similar to the "bash-sources/back"
functionality.  The difference here is that we track git refs instead
of working directories.

We're able to receive callbacks from the git repo on each successful
checkout by placing (a link to) this script in the ".git/hooks" dir
and naming it "post-checkout".

See comment on `__git_nav_on_post_checkout_hook()` for specifics of
how this callback is instrumented.