bashrc: MacOS-proof `sed -i` invocations
On MacOS, if you call `sed` with the `-i` flag, you must specify a
"backup" argument. That is,

    $ sed -i.bak 's/foo/bar/g' baz

Will make write out the original contents of `baz` to `baz.bak` and
the substituted contents to `baz`.

On Linux, this argument is optional.  To make the behavior identical,
we can just set the "backup extension" to be the empty string :^)