record-transactions: Strip out sheets integration
This is simpler to build / maintain and less likely to accidentally
overwrite data if I just output the table to stdout instead of trying
to write directly to the google sheet.

This current version is a bit hacked together and under-generalized
(requires editing source code to process each of the different input
types), but that's ok!