vimrc: Auto-detect *.ts and *.tsx as typescript
For some reason, the maintainers of one of the vim typescript plugins
I use decided to turn this off, which was making all of the typescript
files I worked on to be detected as XML.  This was especially annoying
b/c the AV software we use at work doesn't play nicely with libxml2
(seems to be because it's installed via snap and thus cross-filesystem?
not sure why that would be an issue).