gitconf: Move complicated aliases to standalone scripts
Git will automatically take anything in $PATH that starts with "git-"
and allow you to execute it with the main `git` entrypoint.  For example,
if `git-foo` is on your `$PATH`, then

    $ git foo

is the same as

    $ git-foo

That means we can move a lot of the complexity from the aliases out-of-line
and into their own scripts (for readability).