refactor: Migrate to ansible!
Once everything is installed, the interface is basically the same --
just `dot pull`.  The installation is now just

    $ ./bootstrap

for the first installation and

    $ dot-sync

for further refreshes.

The basic reasoning here is that `ansible` is just much easier to
manage for this kind of task, as it provides good tools out of the
box for idempotency and such :^)

One wrinkle is that we're now installing using a dedicated virtual
environment to manage the installation of `ansible` itself :^)