git-relativize: Add helper to repo-relative paths cwd-relative
When you run `git diff --name-only HEAD^`, you get a list of files
that you touched in the last commit.  However, these files are always
printed with the full path (relative to the root of the repo), even
if you're in a random subdir.

That means, you need to be in the repo root in order to do things
like

    $ git diff --name-only HEAD^ | xargs wc -l

but now you can do it from anywhere with

    $ git diff --name-only HEAD^ | git relativize | xargs wc -l